cleanup

+++ cachemanager-todo.azurewebsites.net/Website/Models/Todo.cs
@@ -1,7 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using Newtonsoft.Json;
 
 namespace Website.Models
@@ -11,7 +9,7 @@ namespace Website.Models
     public class Todo
     {
         [JsonProperty(PropertyName = "id")]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         [JsonProperty(PropertyName = "title")]
         public string Title { get; set; }
